+++ lib/src/generators/emblems/emblem_generator.dart
@@ -3,7 +3,6 @@ import '../../entities/emblems/hsl_color.dart';
 import '../../subtypes/emblems/emblem_type.dart';
 import '../base/batch_generator.dart';
 import '../base/constant_generator.dart ';
-import '../base/future_generator.dart';
 import '../base/generator.dart';
 import '../base/seed_generator.dart';
 import 'icons/emblem_icons_generator.dart';
@@ -35,31 +34,20 @@ class EmblemGenerator implements Generator<Emblem> {
         BatchGenerator(_getBatch(primaryColor, secondaryColor, iconCount));
     generator.seed((_seed + 3) % SeedGenerator.maxSeed);
 
-    return Emblem.fromMap(generator.generate());
+    return Emblem.fromShallowMap(generator.generate());
   }
 
   Map<String, Generator> _getBatch(
           HslColor primaryColor, HslColor secondaryColor, int iconCount) =>
       {
-        "shape": FutureGenerator(
-          _emblemType.getShapeGenerator(),
-          (svg) => svg.toMap(),
-        ),
-        "pattern": FutureGenerator(
-          _emblemType.getPatternGenerator(),
-          (svg) => svg.toMap(),
-        ),
-        "icons": FutureGenerator(
-          EmblemIconsGenerator(_emblemType, iconCount),
-          (icons) => icons.map((e) => e.toMap()).toList(),
-        ),
-        "primaryColor": ConstantGenerator(primaryColor.toMap()),
-        "secondaryColor": ConstantGenerator(secondaryColor.toMap()),
-        "iconsColor": FutureGenerator(
-          _emblemType.getIconsColorGenerator(primaryColor, secondaryColor),
-          (color) => color.toMap(),
-        ),
-        "type": ConstantGenerator(_emblemType.getEmblemType())
+        "shape": _emblemType.getShapeGenerator(),
+        "pattern": _emblemType.getPatternGenerator(),
+        "icons": EmblemIconsGenerator(_emblemType, iconCount),
+        "primaryColor": ConstantGenerator(primaryColor),
+        "secondaryColor": ConstantGenerator(secondaryColor),
+        "iconsColor":
+            _emblemType.getIconsColorGenerator(primaryColor, secondaryColor),
+        "type": ConstantGenerator(_emblemType)
       };
 
   @override